requirements.txt
azure.functions
azure-functions-worker
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
protobuf==3.20.3
requests==2.31.0
six==1.16.0
urllib3==2.0.3
